feat: using visitor pattern
