+++ stylers/Cargo.toml
@@ -25,7 +25,7 @@ clap = ["dep:clap"]
 
 [dependencies]
 glob = "0.3.1"
-syn = { version = "2.0.15", features = ["full"] }
+syn = { version = "2.0.15", features = ["extra-traits", "full", "visit"] }
 stylers_core = "1.0.2"
 stylers_macro = "1.0.2"
 proc-macro2 = "1.0.60"
